Folded in the revised Order Receive Manager

+++ DataLayer/Entity/PurchaseOrder.cs
@@ -32,7 +32,7 @@ namespace DataLayer.Entity
         public DateTime? RecievedDate { get; set; }
         public decimal? SubTotal { get; set; }
         public string Memo { get; set; }
-        public int? OrderState { get; set; }
+        public byte? OrderState { get; set; }
         public DateTime? DateAdded { get; set; }
         public DateTime? ModifiedDate { get; set; }
         public string AddedBy { get; set; }
